fixed autogenerated types for sum and quantity fields in DTOs and Models https://github.com/semashkinvg/Bitmex.NET/projects/3#card-9706627

+++ Bitmex.NET.IntegrationTests/Tests/BitmexApiServiceOrderApiTests.cs
@@ -9,7 +9,7 @@ namespace Bitmex.NET.IntegrationTests.Tests
 	[TestClass]
 	public class BitmexApiServiceOrderApiTests : IntegrationTestsClass<IBitmexApiService>
 	{
-		private double _xbtAvgPrice;
+		private decimal _xbtAvgPrice;
 
 		[ClassCleanup]
 		public static void ClassCleanup()
